* KSoft: Moved some bitwise methods from IntegerMath to the Bits class * KSoft: Changed the bit position lookup tables in Bits to be bytes instead of, the more wasteful, ints. Methods that return looked up values now return bytes, enabling the Ensures contract for >= 0 to be removed. * KSoft.IO.TagElementStreams: Allowed the way Guids are seraialized to be configurable

+++ KSoft/Collections/BitSet.cs
@@ -36,7 +36,7 @@ namespace KSoft.Collections
 		static readonly Bits.VectorIndexInT kVectorIndexInT;
 		static readonly Bits.VectorBitCursorInT kVectorBitCursorInT;
 
-		static readonly Func<TWord, int> kCountZerosForNextBit;
+		static readonly Func<TWord, byte> kCountZerosForNextBit;
 
 		static BitSet()
 		{
@@ -56,8 +56,8 @@ namespace KSoft.Collections
 			// Big:    Bits go from MSB->LSB, so we want to count the 'left most' zeros
 			// Little: Bits go from LSB->MSB, so we want to count the 'right most' zeros
 			kCountZerosForNextBit = Bits.kVectorWordFormat == Shell.EndianFormat.Big
-				? (Func<TWord,int>)Bits.LeadingZerosCount   // Big Endian
-				: (Func<TWord,int>)Bits.TrailingZerosCount; // Little Endian
+				? (Func<TWord,byte>)Bits.LeadingZerosCount   // Big Endian
+				: (Func<TWord,byte>)Bits.TrailingZerosCount; // Little Endian
 #pragma warning restore 0162
 		}
